rename aeroframe to aeroframe_new
