+++ README.md
@@ -173,7 +173,7 @@ There are many different modules available in CEASIOMpy that can be combined to 
 
 #### Structure
 
-- [AeroFrame](./ceasiompy/AeroFrame_new/README.md) :heavy_check_mark:
+- [AeroFrame](./ceasiompy/AeroFrame/README.md) :heavy_check_mark:
 
 #### Data Analysis and Storage
